+++ tests/test_utils.py
@@ -5,7 +5,9 @@ from __future__ import unicode_literals
 import unittest
 
 from skosprovider.utils import (
-    dict_dumper
+    dict_dumper,
+    extract_language,
+    add_lang_to_html
 )
 
 from skosprovider.providers import (
@@ -39,7 +41,7 @@ class DictDumperTest(unittest.TestCase):
             ],
             'sources': [
                 {
-                    'citation': u'Monthy Python. Episode Three: How to recognise different types of trees from quite a long way away.', 
+                    'citation': u'Monthy Python. Episode Three: How to recognise different types of trees from quite a long way away.',
                     'markup': None
                 }
             ],
@@ -49,10 +51,10 @@ class DictDumperTest(unittest.TestCase):
             'member_of': ['3'],
             'subordinate_arrays': [],
             'matches': {
-                'close': [
+                'close': [],
+                'exact': [
                     'http://id.python.org/different/types/of/trees/nr/1/the/larch'
                 ],
-                'exact': [],
                 'related': [],
                 'narrow': [],
                 'broad': []
@@ -79,7 +81,10 @@ class DictDumperTest(unittest.TestCase):
                     'note': 'A different type of tree.', 'markup': None
                 }
             ],
-            'sources': [],
+            'sources': [{
+                'citation': '<span class="author">Bicycle repair man</span>',
+                'markup': 'HTML'
+            }],
             'narrower': [],
             'broader': [],
             'related': [],
@@ -115,7 +120,8 @@ class DictDumperTest(unittest.TestCase):
             'sources': [],
             'members': ['1', '2'],
             'member_of': [],
-            'superordinates': []
+            'superordinates': [],
+            'infer_concept_relations': False
         }
         self.world_dump = {
             'id': '1',
@@ -193,3 +199,41 @@ class DictDumperTest(unittest.TestCase):
         dump = dict_dumper(geo)
         dump2 = dict_dumper(self._get_tree_provider(dict_dumper(geo)))
         self.assertEqual(dump, dump2)
+
+class TestExtractLanguage(object):
+
+    def test_extract_language_nlBE(self):
+        assert 'nl-BE' == extract_language('nl-BE')
+
+    def test_extract_language_None(self):
+        assert 'und' == extract_language(None)
+
+
+class TestHtml(object):
+
+    def test_lang_und(self):
+        assert '' == add_lang_to_html('', 'und')
+        assert '<p></p>' == add_lang_to_html('<p></p>', 'und')
+
+    def test_lang_no_html(self):
+        assert '<div xml:lang="en"></div>' == add_lang_to_html('', 'en')
+
+    def test_no_single_child(self):
+        html = '<p>Paragraph 1</p><p>Paragraph2</p>'
+        assert '<div xml:lang="en"><p>Paragraph 1</p><p>Paragraph2</p></div>' == add_lang_to_html(html, 'en')
+
+    def test_text_node(self):
+        html = 'Something'
+        assert '<div xml:lang="en">Something</div>' == add_lang_to_html(html, 'en')
+
+    def test_single_child_no_attributes(self):
+        html = '<p>Paragraph 1</p>'
+        assert '<p xml:lang="en">Paragraph 1</p>' == add_lang_to_html(html, 'en')
+
+    def test_single_child_already_has_langs(self):
+        html = '<p xml:lang="en">Paragraph 1</p>'
+        assert '<p xml:lang="en">Paragraph 1</p>' == add_lang_to_html(html, 'en')
+
+    def test_single_child_other_attributes(self):
+        html = '<p class="something">Paragraph 1</p>'
+        assert '<p class="something" xml:lang="en">Paragraph 1</p>' == add_lang_to_html(html, 'en')